+++ eslint.config.js
@@ -30,6 +30,7 @@ export default tseslint.config(
       'coverage/',
       '*.js',
       'src/openapi/**',
+      'src/openapi-legacy/**',
       '**/*.d.ts',
       'src/openapi/**/*.d.ts',
     ],